+++ controllers/front_page.js
@@ -1,11 +1,26 @@
 define(['baseController', 'module'], function(BaseController, module) {
     return BaseController.override({
-        title: 'My Node.js app based on Pulsr',
+        title: 'Welcome to Radison Guesthouse in Bishkek, Kyrgyzstan!',
         moduleId: module.id,
         layout: 'front-page',
-        css: ['front-page.less'],
+        css: ['front-page.less', 'front-page2.less'],
         pagelets: [
-            {name: 'ga'}
+            {
+                name: 'content',
+                options: {
+                    layout: 'front-page',
+                    contents: [
+                        'welcome',
+                        'accommodation',
+                        'tours-around-kyrgyzstan',
+                        'wifi-internet',
+                        'location'
+                    ]
+                }
+            },
+            {name: 'navigation'},
+            {name: 'ga'},
+            {name: 'footer-nav'}
         ]
     });
 });