+++ pagelets/content/content.js
@@ -1,13 +1,12 @@
 // content.js
 // Displays the page main content
-define(['pagelet', 'underscore', 'fs', 'path', 'module', 'conf', 'handlebars', 'marked'], function(Pagelet, _, fs, path, module, conf, Handlebars, marked) {
-    var pagelet = new Pagelet();
+define(['pagelet', 'underscore', 'fs', 'path', 'module', 'conf'], function(Pagelet, _, fs, path, module, conf) {
+    var pagelet = new Pagelet;
 
     return _.extend(pagelet, {
         moduleUri: module.uri,
         run: function (display, request) {
-            var _this = this,
-                layoutPath = path.join(conf.dir.pagelets, this.dirName, 'layouts', this.options.layout),
+            var layoutPath = path.join(conf.dir.pagelets, this.dirName, 'layouts', this.options.layout),
                 // if there are no contents specifically set to be displayed
                 // then check the request URL if a specific content URL is being
                 // requested.