+++ pagelets/footer-nav/data.js
@@ -1,0 +1,11 @@
+define(function() {
+    // list of menus. Each item will be converted to <a> link.
+    // It's key will become its href, and value its text.
+    return [
+        {href:'/', text:'Home'},
+        {href: '/accommodation', text: 'Accommodation'},
+        {href: '/location', text: 'Location'},
+        {href: '/tours-around-kyrgyzstan', text: 'Tours around Kyrgyzstan!'},
+        {href: '/contacts', text: 'Contacts'}
+    ];
+});